adapter db order, cashback and product separated
