+++ src/core/types/order.ts
@@ -1,7 +1,14 @@
 import { Product } from './product'
-type ProductOrder = Product & { quantity: number }
+export type ProductOrder = Product & { quantity: number }
 
 export type Order = {
   orderValue: number
   productList: ProductOrder[]
 }
+
+export type OrderBody = {
+  id: string
+  quantity: number
+}
+
+export type ProductsDatabase = OrderBody & Product
