+++ src/core/use-cases/order/order.test.ts
@@ -1,22 +1,16 @@
 import * as TE from 'fp-ts/TaskEither'
 import { pipe } from 'fp-ts/function'
-import { Order } from '@/core/types/order'
+import { ProductOrder } from '@/core/types/order'
 import { createOrder, SaveOrder } from './create-order'
 
-const order: Order = {
-  orderValue: 100,
-  productList: [
-    { name: 'product1', price: 100, quantity: 10 },
-    { name: 'product2', price: 200, quantity: 1 },
-  ],
-}
-const orderProductListEmpty: Order = {
-  orderValue: 0,
-  productList: [],
-}
+const order: ProductOrder[] = [
+  { name: 'product1', price: 100, quantity: 10 },
+  { name: 'product2', price: 200, quantity: 1 },
+]
+const orderProductListEmpty: ProductOrder[] = []
 
 const saveOrder: SaveOrder = async (order) => {
-  return `Pedido cadastrado com sucesso! ${order.orderValue}`
+  return `Pedido cadastrado com sucesso! ${order.length}`
 }
 const saveOrderError: SaveOrder = async () => {
   throw new Error('Database Error!')
@@ -26,7 +20,7 @@ it('Deve criar um pedido', async () => {
   return pipe(
     order,
     createOrder(saveOrder),
-    TE.map((newOrder) => expect(newOrder).toBe(`Pedido cadastrado com sucesso! ${order.orderValue}`)),
+    TE.map((newOrder) => expect(newOrder).toBe(`Pedido cadastrado com sucesso! ${order.length}`)),
   )()
 })
